Create an example and refine some functions

+++ inc/MQTT_osal.h
@@ -1,6 +1,6 @@
 #ifndef __MQTT_OSAL_H__
 #define __MQTT_OSAL_H__
 
-unsigned long MQTT_timestamp();
+unsigned long long MQTT_timestamp();
 
 #endif
