+++ src/MQTT_osal.c
@@ -4,13 +4,13 @@
 #endif
 
 // Get timestamp in milliseconds
-unsigned long MQTT_timestamp()
+unsigned long long MQTT_timestamp()
 {
     unsigned long timestamp;
 #if defined(_WIN32) || defined(_WIN64)
     FILETIME file_time;
     GetSystemTimeAsFileTime(&file_time);
-    timestamp = ((unsigned long)file_time.dwHighDateTime << 32) + (unsigned long)file_time.dwLowDateTime;
+    timestamp = ((unsigned long long )file_time.dwHighDateTime << 32) + (unsigned long long)file_time.dwLowDateTime;
 #endif
     return timestamp;
 }
